@@ -3,21 +3,21 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
-		"@babel/parser": "^7.12.7",
-		"@material-ui/core": "^4.11.0",
-		"@material-ui/lab": "^4.0.0-alpha.56",
-		"@testing-library/jest-dom": "^4.2.4",
-		"@testing-library/react": "^9.3.2",
-		"@testing-library/user-event": "^7.1.2",
-		"@uiw/react-md-editor": "^2.0.3",
-		"axios": "^0.21.1",
+		"@babel/parser": "^7.15.7",
+		"@material-ui/core": "^4.12.3",
+		"@material-ui/lab": "^4.0.0-alpha.60",
+		"@testing-library/jest-dom": "^5.14.1",
+		"@testing-library/react": "^12.1.0",
+		"@testing-library/user-event": "^13.2.1",
+		"@uiw/react-md-editor": "^3.6.3",
+		"axios": "^0.21.4",
 		"balloon-css": "^1.2.0",
 		"prop-types": "^15.7.2",
-		"react": "^16.13.1",
-		"react-dom": "^16.13.1",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2",
 		"react-loading": "^2.0.3",
-		"react-router-dom": "^5.2.0",
-		"react-scripts": "3.4.1"
+		"react-router-dom": "^5.3.0",
+		"react-scripts": "4.0.3"
 	},
 	"scripts": {
 		"start": "react-scripts start",
